chore: regroup and rename form fields components
